Separate multi-version bitmap functions from generic code
