+++ libass/Makefile_library.am
@@ -25,9 +25,13 @@ libass_libass_la_SOURCES = \
     libass/ass_outline.h libass/ass_outline.c \
     libass/ass_drawing.h libass/ass_drawing.c \
     libass/ass_bitmap.h libass/ass_bitmap.c libass/ass_blur.c \
-    libass/ass_rasterizer.h libass/ass_rasterizer.c libass/ass_rasterizer_c.c \
+    libass/ass_rasterizer.h libass/ass_rasterizer.c \
     libass/ass_render.h libass/ass_render.c libass/ass_render_api.c \
     libass/ass_bitmap_engine.h libass/ass_bitmap_engine.c \
+    libass/c/c_rasterizer.c \
+    libass/c/c_blend_bitmaps.c \
+    libass/c/c_be_blur.c \
+    libass/c/c_blur.c \
     libass/wyhash.h
 
 if ASM
@@ -56,6 +60,7 @@ if CORETEXT
 libass_libass_la_SOURCES += libass/ass_coretext.h libass/ass_coretext.c
 endif
 
+libass_libass_la_CPPFLAGS = -I$(top_srcdir)/libass
 libass_libass_la_LDFLAGS = -no-undefined -version-info $(LIBASS_LT_CURRENT):$(LIBASS_LT_REVISION):$(LIBASS_LT_AGE)
 libass_libass_la_LDFLAGS += -export-symbols $(top_srcdir)/libass/libass.sym
