+++ libass/ass_bitmap.c
@@ -258,120 +258,3 @@ void ass_shift_bitmap(Bitmap *bm, int shift_x, int shift_y)
             }
         }
 }
-
-/**
- * \brief Blur with [[1,2,1], [2,4,2], [1,2,1]] kernel
- * This blur is the same as the one employed by vsfilter.
- * Pure C implementation.
- */
-void ass_be_blur_c(uint8_t *buf, intptr_t stride,
-                   intptr_t width, intptr_t height, uint16_t *tmp)
-{
-    uint16_t *col_pix_buf = tmp;
-    uint16_t *col_sum_buf = tmp + width;
-    unsigned x, y, old_pix, old_sum, temp1, temp2;
-    uint8_t *src, *dst;
-    y = 0;
-
-    {
-        src=buf+y*stride;
-
-        x = 1;
-        old_pix = src[x-1];
-        old_sum = old_pix;
-        for ( ; x < width; x++) {
-            temp1 = src[x];
-            temp2 = old_pix + temp1;
-            old_pix = temp1;
-            temp1 = old_sum + temp2;
-            old_sum = temp2;
-            col_pix_buf[x-1] = temp1;
-            col_sum_buf[x-1] = temp1;
-        }
-        temp1 = old_sum + old_pix;
-        col_pix_buf[x-1] = temp1;
-        col_sum_buf[x-1] = temp1;
-    }
-
-    for (y++; y < height; y++) {
-        src=buf+y*stride;
-        dst=buf+(y-1)*stride;
-
-        x = 1;
-        old_pix = src[x-1];
-        old_sum = old_pix;
-        for ( ; x < width; x++) {
-            temp1 = src[x];
-            temp2 = old_pix + temp1;
-            old_pix = temp1;
-            temp1 = old_sum + temp2;
-            old_sum = temp2;
-
-            temp2 = col_pix_buf[x-1] + temp1;
-            col_pix_buf[x-1] = temp1;
-            dst[x-1] = (col_sum_buf[x-1] + temp2) >> 4;
-            col_sum_buf[x-1] = temp2;
-        }
-        temp1 = old_sum + old_pix;
-        temp2 = col_pix_buf[x-1] + temp1;
-        col_pix_buf[x-1] = temp1;
-        dst[x-1] = (col_sum_buf[x-1] + temp2) >> 4;
-        col_sum_buf[x-1] = temp2;
-    }
-
-    {
-        dst=buf+(y-1)*stride;
-        for (x = 0; x < width; x++)
-            dst[x] = (col_sum_buf[x] + col_pix_buf[x]) >> 4;
-    }
-}
-
-/**
- * \brief Add two bitmaps together at a given position
- * Uses additive blending, clipped to [0,255]. Pure C implementation.
- */
-void ass_add_bitmaps_c(uint8_t *dst, intptr_t dst_stride,
-                       uint8_t *src, intptr_t src_stride,
-                       intptr_t width, intptr_t height)
-{
-    unsigned out;
-    uint8_t* end = dst + dst_stride * height;
-    while (dst < end) {
-        for (unsigned j = 0; j < width; ++j) {
-            out = dst[j] + src[j];
-            dst[j] = FFMIN(out, 255);
-        }
-        dst += dst_stride;
-        src += src_stride;
-    }
-}
-
-void ass_imul_bitmaps_c(uint8_t *dst, intptr_t dst_stride,
-                        uint8_t *src, intptr_t src_stride,
-                        intptr_t width, intptr_t height)
-{
-    uint8_t* end = dst + dst_stride * height;
-    while (dst < end) {
-        for (unsigned j = 0; j < width; ++j) {
-            dst[j] = (dst[j] * (255 - src[j]) + 255) >> 8;
-        }
-        dst += dst_stride;
-        src += src_stride;
-    }
-}
-
-void ass_mul_bitmaps_c(uint8_t *dst, intptr_t dst_stride,
-                       uint8_t *src1, intptr_t src1_stride,
-                       uint8_t *src2, intptr_t src2_stride,
-                       intptr_t width, intptr_t height)
-{
-    uint8_t* end = src1 + src1_stride * height;
-    while (src1 < end) {
-        for (unsigned x = 0; x < width; ++x) {
-            dst[x] = (src1[x] * src2[x] + 255) >> 8;
-        }
-        dst  += dst_stride;
-        src1 += src1_stride;
-        src2 += src2_stride;
-    }
-}
